requirements.txt
absl-py==0.7.1
appdirs==1.4.3
aspy.yaml==1.3.0
astor==0.8.0
attrs==19.1.0
black==19.3b0
cfgv==2.0.0
Click==7.0
cycler==0.10.0
-e git+https://github.com/TimGuiteDiamond/topaz3.git#egg=topaz3
entrypoints==0.3
flake8==3.7.7
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
identify==1.4.5
importlib-metadata==0.18
importlib-resources==1.0.2
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
logconfig==0.4.0
logutils==0.3.5
Markdown==3.1.1
matplotlib==3.1.0
mccabe==0.6.1
nodeenv==1.3.3
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
pre-commit==1.17.0
procrunner==1.0.2
protobuf==3.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
toml==0.10.0
virtualenv==16.6.1
Werkzeug==0.15.4
zipp==0.5.1
